% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_df_for_stata.R
\name{assign_variable_labels}
\alias{assign_variable_labels}
\title{Assign variable labels}
\usage{
assign_variable_labels(
  dataset,
  dictionary,
  ignore_columns,
  var_label,
  var_colname,
  varnames
)
}
\arguments{
\item{dataset}{Any data frame with categorical variables coded as numbers.}

\item{dictionary}{A data frame that serves as a dictionary,
mapping each numeric code to its corresponding category string.}

\item{ignore_columns}{You don't want to modify the columns that
serves as the id of the data or numeric variables
that are not categorical variables. You can specify them here.}

\item{var_label}{Name of the dictionary column containing the
dataset variable labels.
To understand the dictionary format,
refer to the documentation for the `dict_df_var_label` dataset.}

\item{var_colname}{Name of the dictionary column containing the
dataset variable column names.
To understand the dictionary format,
refer to the documentation for the `dict_df_var_label` dataset.}

\item{varnames}{Variable names}
}
\value{
The original dataset with variable labels
}
\description{
Assign variable labels
}
\examples{

library(dplyr)

varnames <- dict_df_var_label |>
select(variable) |>
filter(
   variable \%in\% colnames(df)) |> # I remove the variables that are not in the data
 unique() |>
 pull()

assign_variable_labels(dataset = df,
                    dictionary = dict_df_var_label,
                    ignore_columns = c("id", "age"),
                    var_label = var_label ,
                    var_colname = variable,
                    varnames = varnames)

}
