% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-df_missing-dict_df_missing.R
\docType{data}
\name{df_missing}
\alias{df_missing}
\title{Sample dataset with missing values}
\format{
## `df_missing`
A data frame with 10 rows and 5 columns:
\describe{
  \item{id}{Unique identifier for each record.}
  \item{gender}{Gender of the individual (factor: "Male", "Female").}
  \item{age}{Age.}
  \item{blood_pressure_category}{Blood pressure category (numeric; includes codes for missing values).}
  \item{glucose_level}{Glucose level category (numeric).}
  \item{cholesterol_level}{Cholesterol level category (numeric; includes codes for missing values).}
}
}
\source{
Generated as an example dataset.
}
\usage{
df_missing
}
\description{
A sample dataset (`df_missing`) containing missing values in categorical numeric variables and an associated dictionary (`dict_df_missing`) that defines category labels, including indicators for missing and special cases.
}
\keyword{datasets}
