% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_df_for_stata.R
\name{label_df_for_STATA}
\alias{label_df_for_STATA}
\title{Add variable and value labels compatible with STATA
to all variables in the dataset}
\usage{
label_df_for_STATA(
  dataset,
  dictionary,
  ignore_columns,
  var_label,
  var_colname,
  num_colname,
  str_colname
)
}
\arguments{
\item{dataset}{Any data frame with categorical variables coded as numbers.}

\item{dictionary}{A data frame that serves as a dictionary,
mapping each numeric code to its corresponding category string.}

\item{ignore_columns}{You don't want to modify the columns that
serves as the id of the data or numeric variables
that are not categorical variables. You can specify them here.}

\item{var_label}{Name of the dictionary column containing the
dataset variable labels.
To understand the dictionary format,
refer to the documentation for the `dict_df_var_label` dataset.}

\item{var_colname}{Name of the dictionary column containing the
dataset variable column names.
To understand the dictionary format,
refer to the documentation for the `dict_df_var_label` dataset.}

\item{num_colname}{Name of the dictionary column containing
the numeric codes for the categories.
To understand the dictionary format,
refer to the documentation for the `dict_df_var_label` dataset.}

\item{str_colname}{Name of the dictionary column containing the
labels or categories as character strings.
To understand the dictionary format,
refer to the documentation for the `dict_df_var_label` dataset.}
}
\value{
The original dataset with variable and code labels
}
\description{
`label_df_for_STATA()` allows you to add labels to variable names
and categorical numeric variables observations (called here "values").
The function applies labels to the entire dataset
by reading information from a dictionary.
For an example of the required dictionary format,
see the dataset [dict_df_var_label].
}
\details{
For variable name labels,
the function `label_df_for_STATA()` fills in missing labels
from the dictionary with empty labels.
This means that the function won’t fail due to missing labels.

For value labels,
you can specify the numeric variables (e.g., id and age)
that don't need labeling as they are not categorical.
Note that STATA only reads labels for numeric variables.
}
\examples{

label_df_for_STATA(dataset = df,
                    dictionary = dict_df_var_label,
                    ignore_columns = c("id", "age"),
                    var_label = var_label ,
                    var_colname = variable,
                    num_colname = level_num,
                    str_colname = level_str)
}
