% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing_with_na.R
\name{replace_missing_with_na}
\alias{replace_missing_with_na}
\title{Replaces in a data frame values coded as "missing" in the dictionary with NA}
\usage{
replace_missing_with_na(
  dataset,
  ignore_columns,
  dictionary,
  var_colname,
  num_colname,
  missing_colname
)
}
\arguments{
\item{dataset}{Any data frame with categorical variables coded as numbers.
To see an example you can refer to the documentation
for the data frame [df_missing].}

\item{ignore_columns}{You don't want to modify the columns that
serves as the id of the data or numeric variables
that are not categorical variables.
You can specify them here.}

\item{dictionary}{A data frame that serves as a dictionary,
mapping each numeric code to its corresponding category string.
This dictionary includes a column that indicates
whether a number represents a missing value (1)
or a reported value (0).
In CLSA, this column is included by default and
is called `missing`.
If you wish to use this function with another dataset,
you can create this column.
To understand what the dictionary looks like,
you can refer to the documentation
for the example dictionary [dict_df_missing].}

\item{var_colname}{Name of the dictionary column containing the
dataset variable names.
To understand the dictionary format,
refer to the documentation for the `dict_df_missing` dataset.}

\item{num_colname}{Name of the dictionary column containing
the numeric codes for the categories.
To understand the dictionary format,
refer to the documentation for the `dict_df_missing` dataset.}

\item{missing_colname}{Name of the dictionary column that
is a binary indicator where a value of 1 represents a missing
or unreported observation, and a value of 0
represents a reported observation.
To understand the dictionary format,
refer to the documentation for the `dict_df_missing` dataset.}
}
\value{
The original dataset with NAs in all missing values for columns with
 categorical numeric variables.
}
\description{
In the dictionary,
there are often specific numeric codes
used to represent missing observations
(for example, a common one is -99).

Although it’s usually the same value,
there may be more than one.
For example, the Canadian Longitudinal Study in Aging (CLSA)
distinguishes between -9999 for missing values
and -8888 for cases where the respondent did not answer.

The function `replace_missing_with_na` assumes
the presence of a column (called `missing` in CLSA dictionary)
that codes 0 if the value is present and 1 if it is absent.

Notice that it would be relatively straightforward
to remove rows where `missing == 1` to remove missing data.
The purpose of `replace_with_na()` is to convert
coded values to `NA` to accurately estimate
the total amount of missing data in the database,
regardless of the reason.
}
\examples{

replace_missing_with_na(dataset = df_missing,
                        ignore_columns = c(id, age),
                        dictionary = dict_df_missing,
                        var_colname = variable,
                        num_colname = level_num,
                        missing_colname = missing)
}
