% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_df_for_stata.R
\name{label_df_for_STATA}
\alias{label_df_for_STATA}
\title{Add labels readable by STATA to all the columns automatically}
\usage{
label_df_for_STATA(
  dataset,
  dictionary,
  ignore_columns,
  var_label,
  var_colname,
  num_colname,
  str_colname
)
}
\arguments{
\item{dataset}{Any data frame with categorical variables coded as numbers.}

\item{dictionary}{A data frame that serves as a dictionary,
mapping each numeric code to its corresponding category string.}

\item{ignore_columns}{You don't want to modify the columns that
serves as the id of the data or numeric variables
that are not categorical variables. You can specify them here.}

\item{var_label}{Name of the dictionary column containing the
dataset variable labels.
To understand the dictionary format,
refer to the documentation for the `dict_df_var_label` dataset.}

\item{var_colname}{Name of the dictionary column containing the
dataset variable column names.
To understand the dictionary format,
refer to the documentation for the `dict_df_var_label` dataset.}

\item{num_colname}{Name of the dictionary column containing
the numeric codes for the categories.
To understand the dictionary format,
refer to the documentation for the `dict_df_var_label` dataset.}

\item{str_colname}{Name of the dictionary column containing the
labels or categories as character strings.
To understand the dictionary format,
refer to the documentation for the `dict_df_var_label` dataset.}
}
\value{
The original dataset with variable and code labels
}
\description{
There are 4 elements that are read in STATA
The variable name
The variable label
The value code
The value label
}
\details{
Note: is important that you have labels for all the variables
}
\examples{

label_df_for_STATA(dataset = df,
                    dictionary = dict_df_var_label,
                    ignore_columns = c("id", "age"),
                    var_label = var_label ,
                    var_colname = variable,
                    num_colname = level_num,
                    str_colname = level_str)
}
